refactor: drag 관련 로직을 hook으로 분리
