+++ src/components/ResizableContainer/index.tsx
@@ -1,5 +1,5 @@
-import { getPosition, mouseDrag } from '@/utils/mouseDrag';
-import { HTMLAttributes, useState, RefObject, ReactElement, useRef } from 'react';
+import useDraggable from '@/components/ResizableContainer/useDraggable';
+import { HTMLAttributes, RefObject, ReactElement } from 'react';
 
 interface Props {
   initialX: number;
@@ -15,127 +15,22 @@ const ResizableContainer = ({
   render,
   ...props
 }: HTMLAttributes<HTMLElement> & Props) => {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const [{ w, h }, setSize] = useState({ w: 200, h: 200 });
-  const [{ x, y }, setPosition] = useState({ x: initialX, y: initialY });
-
-  const { handleMouseDown: handleDragElement } = mouseDrag((moveX, moveY) => {
-    if (boundaryRef.current && containerRef.current) {
-      const containerRect = boundaryRef.current.getBoundingClientRect();
-      const targetElementRect = containerRef.current.getBoundingClientRect();
-      const calculatedX = getPosition(x + moveX, 0, containerRect.width - targetElementRect.width);
-      const calculatedY = getPosition(y + moveY, 0, containerRect.height - targetElementRect.height);
-      setPosition({ x: calculatedX, y: calculatedY });
-    }
-  });
-
-  const { handleMouseDown: handleResizeNorthWest } = mouseDrag((moveX, moveY) => {
-    setSize({
-      w: getPosition(w - moveX, 200, x + w),
-      h: getPosition(h - moveY, 200, y + h),
-    });
-    setPosition({
-      x: getPosition(x + moveX, 0, x + w - 200),
-      y: getPosition(y + moveY, 0, y + h - 200),
-    });
-  });
-
-  const { handleMouseDown: handleResizeNorthEast } = mouseDrag((moveX, moveY) => {
-    if (boundaryRef.current) {
-      const boundaryRect = boundaryRef.current.getBoundingClientRect();
-
-      setSize({
-        w: getPosition(w + moveX, 200, boundaryRect.width - x),
-        h: getPosition(h - moveY, 200, y + h),
-      });
-      setPosition({
-        x,
-        y: getPosition(y + moveY, 0, y + h - 200),
-      });
-    }
-  });
-
-  const { handleMouseDown: handleResizeSouthWest } = mouseDrag((moveX, moveY) => {
-    if (boundaryRef.current) {
-      const boundaryRect = boundaryRef.current.getBoundingClientRect();
-
-      setSize({
-        w: getPosition(w - moveX, 200, x + w),
-        h: getPosition(h + moveY, 200, boundaryRect.height - y),
-      });
-      setPosition({
-        x: getPosition(x + moveX, 0, x + w - 200),
-        y,
-      });
-    }
-  });
-
-  const { handleMouseDown: handleResizeSouthEast } = mouseDrag((moveX, moveY) => {
-    if (boundaryRef.current) {
-      const boundaryRect = boundaryRef.current.getBoundingClientRect();
-
-      setSize({
-        w: getPosition(w + moveX, 200, boundaryRect.width - x),
-        h: getPosition(h + moveY, 200, boundaryRect.height - y),
-      });
-      setPosition({
-        x,
-        y,
-      });
-    }
-  });
-
-  const { handleMouseDown: handleResizeWest } = mouseDrag((moveX) => {
-    setSize({
-      w: getPosition(w - moveX, 200, x + w),
-      h,
-    });
-    setPosition({
-      x: getPosition(x + moveX, 0, x + w - 200),
-      y,
-    });
-  });
-
-  const { handleMouseDown: handleResizeEast } = mouseDrag((moveX) => {
-    if (boundaryRef.current) {
-      const boundaryRect = boundaryRef.current.getBoundingClientRect();
-
-      setSize({
-        w: getPosition(w + moveX, 200, boundaryRect.width - x),
-        h,
-      });
-      setPosition({
-        x,
-        y,
-      });
-    }
-  });
-
-  const { handleMouseDown: handleResizeNorth } = mouseDrag((_, moveY) => {
-    setSize({
-      w,
-      h: getPosition(h - moveY, 200, y + h),
-    });
-    setPosition({
-      x,
-      y: getPosition(y + moveY, 0, y + h - 200),
-    });
-  });
-
-  const { handleMouseDown: handleResizeSouth } = mouseDrag((_, moveY) => {
-    if (boundaryRef.current) {
-      const boundaryRect = boundaryRef.current.getBoundingClientRect();
-
-      setSize({
-        w,
-        h: getPosition(h + moveY, 200, boundaryRect.height - y),
-      });
-      setPosition({
-        x,
-        y,
-      });
-    }
-  });
+  const {
+    x,
+    y,
+    w,
+    h,
+    handleResizeEast,
+    handleResizeNorth,
+    handleResizeNorthWest,
+    handleResizeNorthEast,
+    handleResizeSouth,
+    handleResizeSouthEast,
+    handleResizeSouthWest,
+    handleResizeWest,
+    containerRef,
+    handleDragElement,
+  } = useDraggable(initialX, initialY, boundaryRef);
 
   return (
     <div {...props} ref={containerRef} style={{ width: w, height: h, transform: `translate(${x}px, ${y}px)` }}>
